경기장 외 배경 무시 기능 완료

See merge request !43

+++ graphic_interface.h
@@ -6,7 +6,7 @@
 
 typedef uint16_t PixelData_t;
 
-typedef uint8_t PixelCoordinate_t;
+typedef int16_t PixelCoordinate_t;
 typedef struct {
     PixelCoordinate_t x;
     PixelCoordinate_t y;
